pyproject.toml
[tool.poetry]
name = "drf-course"
version = "0.1.0"
description = ""
authors = ["gennady <topchiev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.25.0"
setuptools = "^70.0.0"
drf-yasg = "^1.21.7"
eventlet = "^0.36.1"
celery = "^5.4.0"
redis = "^5.0.5"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.3.1"
requests = "^2.32.3"
coverage = "^7.5.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
black = "^24.4.2"

[tool.black]
line-length = 120
extend-exclude = """
    (
    .*migrations/
    | .*tests/
    )
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
